+++ eliot/tests/test_action.py
@@ -174,7 +174,7 @@ class ActionTests(TestCase):
         """
         serializers = ActionType("sys:thename",
                                  [Field("key", lambda x: x, "")],
-                                 [], [], "")._serializers
+                                 [], "")._serializers
         class Logger(list):
             def write(self, msg, serializer):
                 self.append(serializer)
@@ -359,7 +359,7 @@ class ActionTests(TestCase):
         serializers = ActionType("sys:thename",
                                  [],
                                  [Field("key", lambda x: x, "")],
-                                 [], "")._serializers
+                                 "")._serializers
         class Logger(list):
             def write(self, msg, serializer):
                 self.append(serializer)
@@ -374,7 +374,7 @@ class ActionTests(TestCase):
         L{Action.finish} with an exception passes the failure
         L{eliot._validation._MessageSerializer} to the message it creates.
         """
-        serializers = ActionType("sys:thename", [], [],
+        serializers = ActionType("sys:thename", [],
                                  [Field("key", lambda x: x, "")],
                                  "")._serializers
         class Logger(list):
@@ -484,56 +484,6 @@ class ActionTests(TestCase):
                              {"x": 1, "y": 2, "z": 3})
 
 
-    def test_addSuccessFieldsIgnoresFailure(self):
-        """
-        On a successful finish, L{Action.__exit__} ignores fields from
-        L{Action.addFailureFields}.
-        """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", "/1/", "sys:me")
-        with action as act:
-            act.addFailureFields(z=3)
-        self.assertNotIn("z", logger.messages[0])
-
-
-    def test_addFailureFields(self):
-        """
-        On an failed finish, L{Action.__exit__} adds fields from
-        L{Action.addFailureFields} to the result message.
-        """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", "/1/", "sys:me")
-        try:
-            with action as act:
-                act.addFailureFields(x=1, y=2)
-                act.addFailureFields(z=3)
-                raise RuntimeError()
-        except RuntimeError:
-            pass
-        else:
-            self.fail("No exception")
-        assertContainsFields(self, logger.messages[0],
-                             {"x": 1, "y": 2, "z": 3})
-
-
-    def test_addFailureFieldsIgnoresSuccess(self):
-        """
-        On a successful finish, L{Action.__exit__} ignores fields from
-        L{Action.addFailureFields}.
-        """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", "/1/", "sys:me")
-        try:
-            with action as act:
-                act.addSuccessFields(z=3)
-                raise RuntimeError()
-        except RuntimeError:
-            pass
-        else:
-            self.fail("No exception")
-        self.assertNotIn("z", logger.messages[0])
-
-
     def test_incrementMessageCounter(self):
         """
         Each call to L{Action._incrementMessageCounter} increments a counter.
@@ -617,13 +567,11 @@ class TwistedActionTests(TestCase):
     def test_finishAfterFailure(self):
         """
         When the L{Deferred} passed to L{Action.finishAfter} fires
-        with an exception, a finish message is logged with any failure bindings
-        added.
+        with an exception, a finish message is logged.
         """
         d = Deferred()
         logger = MemoryLogger()
         action = Action(logger, "uuid", "/1/", "sys:me")
-        action.addFailureFields(x=2)
         action.finishAfter(d)
         exception = RuntimeError("because")
         d.errback(exception)
@@ -632,7 +580,6 @@ class TwistedActionTests(TestCase):
                               "task_level": "/1/",
                               "action_type": "sys:me",
                               "action_status": "failed",
-                              "x": 2,
                               "reason": "because",
                               "exception":
                               "%s.RuntimeError" % (RuntimeError.__module__,)})