+++ docs/source/news.rst
@@ -4,6 +4,10 @@ What's New
 0.4.0
 ^^^^^
 
+Incompatible changes from v0.3:
+
+* ``ActionType`` no longer supports defining additional failure fields, and therefore accepts one argument less.
+
 Features:
 
 * Added support for Python 3.3.